+++ R/modify_source_note.R
@@ -21,11 +21,13 @@
 #'
 #' @return gtsummary object
 #' @name modify_source_note
+#' @seealso [Footnotes vs Source Notes vs Abbreviations](https://www.danieldsjoberg.com/gtsummary/articles/modify-functions.html)
 #'
 #' @examplesIf identical(Sys.getenv("NOT_CRAN"), "true") || identical(Sys.getenv("IN_PKGDOWN"), "true")
 #' # Example 1 ----------------------------------
 #' tbl <- tbl_summary(trial, include = c(marker, grade), missing = "no") |>
 #'   modify_source_note("Results as of June 26, 2015")
+#' tbl
 #'
 #' # Example 2 ----------------------------------
 #' remove_source_note(tbl, source_note_id = 1)