+++ NEWS.md
@@ -1,10 +1,36 @@
 # gtsummary (development version)
 
+# gtsummary 2.5.0
+
+### New Features and Functions
+
+* Added AI chat bot to homepage, powered by `kapa.ai`.
+
+* Added functions `tbl_ard_strata()` and `tbl_ard_strata2()`. (#1852)
+
+### Other Updates
+
+* Added additional messaging to `tbl_merge()` when tables _may not_ merge properly. (#2348)
+
+* When a data frame is passed in the `tbl_summary(percent)` argument, the header counts are now tabulated using this data frame. (#2322)
+
+* Added `head()` and `tail()` S3 methods for `'gtsummary'` objects. (#2335) 
+
 * Updated `tbl_ard_hierarchical()` to pre-process non-standard hierarchical ARDs so sorting can be applied. (#2308)
 
 * Added theme element `"assign_summary_type-arg:cat_threshold"` for greater control over default summary types.
 
-* When a data frame is passed in the `tbl_summary(percent)` argument, the headers are now tabulated with this data frame. (#2322)
+* Added the `"add_overall.tbl_svysummary-arg:col_label"` theme element to control the default value for `add_overall.tbl_svysummary(col_label)` argument. (#2354)
+
+* Added theme elements `add_ci.tbl_summary-arg:method`, and `add_ci.tbl_svysummary-arg:method` to control the default value for `add_ci.tbl_summary(method)` / `add_ci.tbl_svysummary(method)` argument. (#2356)
+
+### Bug Fixes
+
+* Fixed bug in `tbl_strata_nested_stack()` causing incorrect column indentation in non-standard tables. (#2272)
+
+* Fixed bug in `sort_hierarchical()` causing variable labels to disappear from table when not all variables are in `include`. (#2332)
+
+* Corrected handling in `tbl_summary()` of character vectors that contains `"NULL"` values. (#2345)
 
 # gtsummary 2.4.0